fix: huge refactoring of the Dashboard
